fix: services should have destination as well

+++ app/Actions/Service/StopService.php
@@ -20,7 +20,9 @@ class StopService
             instant_remote_process(["docker rm -f {$db->name}-{$service->uuid}"], $service->server);
             $db->update(['status' => 'exited']);
         }
-        instant_remote_process(["docker network disconnect {$service->uuid} coolify-proxy 2>/dev/null"], $service->server, false);
-        instant_remote_process(["docker network rm {$service->uuid} 2>/dev/null"], $service->server, false);
+        if (is_null($service->destination)) {
+            instant_remote_process(["docker network disconnect {$service->uuid} coolify-proxy 2>/dev/null"], $service->server, false);
+            instant_remote_process(["docker network rm {$service->uuid} 2>/dev/null"], $service->server, false);
+        }
     }
 }
